backend/requirements.txt
fastapi
uvicorn
prisma
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
email-validator
hashids
redis
httpx
aiosmtplib
